chore: reorganized package structure for game save components
