+++ lsadf_core/src/test/java/com/lsadf/core/unit/services/StageServiceTests.java
@@ -25,7 +25,7 @@ import static org.mockito.Mockito.when;
 import com.lsadf.core.application.game.save.stage.StageCachePort;
 import com.lsadf.core.application.game.save.stage.StageRepositoryPort;
 import com.lsadf.core.application.game.save.stage.StageService;
-import com.lsadf.core.application.game.save.stage.StageServiceImpl;
+import com.lsadf.core.application.game.save.stage.impl.StageServiceImpl;
 import com.lsadf.core.domain.game.save.stage.Stage;
 import com.lsadf.core.infra.exception.http.NotFoundException;
 import com.lsadf.core.infra.valkey.cache.service.CacheService;
